Updated the code
The output for user is now correct, still need to do some other tuning

+++ User.h
@@ -22,7 +22,7 @@ int findItemIndex(int code){
 class User{
     private:
         static double totalSales;
-        static double totalItems;
+        static int totalItems;
         static int totalTransactions;
 
     protected:
